add to vercel

+++ app/api/shifts/route.js
@@ -3,12 +3,17 @@ import { getServerSession } from 'next-auth/next';
 import prisma from '@/lib/prisma';
 import { authOptions } from '@/lib/authOptions';
 import { Resend } from 'resend';
-import { verificationEmail } from '@/helpers/utils';
+import {
+  checkUnavailabilityConflict,
+  verificationEmail,
+} from '@/helpers/utils';
+import { getNotification, getShiftCreationNotif } from '@/lib/notifications';
 
 // Initialize Resend
 const resend = new Resend(process.env.RESEND_API_KEY);
 
 // GET all shifts
+
 export async function GET() {
   try {
     const session = await getServerSession(authOptions);
@@ -17,31 +22,34 @@ export async function GET() {
       return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
     }
 
-    // // Role-based authorization
-    // const userRole = session.user.role;
-    // const allowedRoles = ['admin'];
+    const isAuthorised =
+      session.user.role === 'admin' || session.user.role === 'owner';
+    const userId = session.user.id;
 
-    // if (!allowedRoles.includes(userRole)) {
-    //   return res.status(403).json({
-    //     error: 'You do not have permission to perform this action.',
-    //   });
-    // }
-
+    // Different queries based on role
     const shifts = await prisma.shift.findMany({
-      where: {
-        // Optional: filter shifts based on user role or other criteria
-        // userId: session.user.id
-      },
+      where: isAuthorised
+        ? {} // Admin sees all shifts
+        : { assignedToUser: { id: userId } }, // Users see only their assigned shifts
       include: {
         assignedToUser: true,
         createdByUser: true,
+        exchangeRequests: {
+          where: {
+            status: 'PENDING',
+          },
+          select: {
+            id: true,
+            status: true,
+            // ... any other fields you need
+          },
+        },
       },
       orderBy: {
         start: 'asc',
       },
     });
 
-    // console.log('List of shift🔥:', shifts);
     return NextResponse.json(shifts);
   } catch (error) {
     console.error('Shifts fetch error:', error);
@@ -84,7 +92,42 @@ export async function POST(request) {
       sendNotification = false,
     } = await request.json();
 
-    // First, fetch the assigned user to get their name for the title
+    // Combine date and times to create full datetime objects
+    const shiftStart = new Date(`${date}T${startTime}`);
+    const shiftEnd = new Date(`${date}T${endTime}`);
+
+    // Fetch user with unavailabilities
+    const userWithUnavailabilities = await prisma.user.findUnique({
+      where: { id: employee },
+      select: {
+        unavailabilities: true,
+        firstName: true,
+        lastName: true,
+        email: true,
+      },
+    });
+
+    if (userWithUnavailabilities?.unavailabilities.length > 0) {
+      const hasUnavailabilityConflict = checkUnavailabilityConflict(
+        shiftStart,
+        shiftEnd,
+        userWithUnavailabilities.unavailabilities
+      );
+
+      console.log('User with unavailabilities🩸', hasUnavailabilityConflict);
+
+      if (hasUnavailabilityConflict) {
+        return NextResponse.json(
+          {
+            error: `${userWithUnavailabilities.firstName} ${userWithUnavailabilities.lastName} is unavailable during this time.`,
+            unavailabilities: userWithUnavailabilities.unavailabilities,
+          },
+          { status: 400 }
+        );
+      }
+    }
+
+    // Fetch the assigned user to get their name for the title
     const assignedUser = await prisma.user.findUnique({
       where: { id: employee }, // employee should now be the user ID
       select: {
@@ -129,10 +172,32 @@ export async function POST(request) {
             firstName: true,
             lastName: true,
             email: true,
+            avatarColor: true,
           },
         },
       },
     });
+
+    const message = getNotification({
+      firstName: `${shift.createdByUser.firstName} ${shift.createdByUser.lastName}`,
+      firstNameAvatarColour: `${shift.createdByUser.avatarColor}`,
+      action: `assigned a shift to you`,
+    });
+
+    // console.log('message🩸🪗', message);
+    // console.log('shift.createdBy💵', shift.createdByUser);
+
+    // Create notifications for relevant users
+    const notification = await prisma.notification.create({
+      data: {
+        userId: shift.assignedToUserId,
+        message: message.plainText,
+        messageData: JSON.stringify(message.data),
+        type: 'TASK_ASSIGNED',
+        link: '/user/my-schedule',
+      },
+    });
+    console.log('Notification created 💵🥇', notification);
 
     // Send email notification if sendNotification is true
     if (sendNotification) {
@@ -178,8 +243,8 @@ export async function POST(request) {
                 : ''
             }
             <p><strong>Assigned by:</strong> ${shift.createdByUser.firstName} ${
-            shift.createdByUser.lastName
-          }</p>
+              shift.createdByUser.lastName
+            }</p>
             ${
               repeatShift
                 ? `<p><strong>This shift repeats:</strong> Every ${repeatFrequency}</p>`
@@ -207,66 +272,6 @@ export async function POST(request) {
     );
   }
 }
-
-// Edit shift
-// export async function PUT(request) {
-//   try {
-//     const session = await getServerSession(authOptions);
-
-//     if (!session) {
-//       return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
-//     }
-
-//     // Fetch user to check role
-//     const user = await prisma.user.findUnique({
-//       where: { email: session.user.email },
-//     });
-
-//     if (user?.role !== 'admin') {
-//       return NextResponse.json({ error: 'Forbidden' }, { status: 403 });
-//     }
-
-//     const {
-//       id,
-//       start,
-//       end,
-//       title,
-//       description,
-//       userId,
-//       breaks,
-//       shiftType,
-//       repeatShift,
-//       repeatFrequency,
-//       sendNotification,
-//     } = await request.json();
-
-//     const shift = await prisma.shift.update({
-//       where: { id },
-//       data: {
-//         // start: new Date(start),
-//         start: new Date(`${date}T${startTime}`),
-//         end: new Date(`${date}T${endTime}`),
-//         // end: new Date(end),
-//         title,
-//         description,
-//         userId,
-//         breaks,
-//         shiftType,
-//         repeatShift,
-//         repeatFrequency,
-//         sendNotification,
-//       },
-//     });
-
-//     return NextResponse.json(shift);
-//   } catch (error) {
-//     console.error('Shift update error:', error);
-//     return NextResponse.json(
-//       { error: 'Failed to update shift' },
-//       { status: 500 }
-//     );
-//   }
-// }
 
 export async function PUT(request) {
   try {
@@ -363,6 +368,16 @@ export async function DELETE(request) {
     const { searchParams } = new URL(request.url);
     const id = searchParams.get('id');
 
+    const shift = await prisma.shift.findUnique({
+      where: {
+        id,
+        createdByUserId: session.user.id,
+      },
+    });
+
+    if (!shift) {
+      return NextResponse.json({ error: 'Shift not found' }, { status: 404 });
+    }
     await prisma.shift.delete({
       where: { id },
     });
